Fixes & Improve delete file from P.M
- this fixes , no button for mediainfo and ss in file msg
- also added DELETE_FILES_FROM_PMS to control delete file after 10mins from user's pm

+++ bot.py
@@ -70,28 +70,23 @@ async def _start(event):
     if msg_id:
         if msg_id.isdigit():
             msg = await bot.get_messages(Var.BACKUP_CHANNEL, ids=int(msg_id))
-            sent_msg = await event.reply(msg.text, file=msg.media)
-            notice = await event.reply(
-                "⚠️ **Important Notice:**\n\n```This file will be automatically deleted after 10 minutes.```\n__Please save or forward it immediately.__"
-            )
-            asyncio.create_task(bot.delete_after([notice, sent_msg]))
+            sent_msg = await event.reply(msg)
+            if Var.DELETE_FILES_FROM_PMS:
+                notice = await sent_msg.reply(
+                    "__This file will be automatically deleted after 10 minutes.\nPlease save or forward it immediately.__"
+                )
+                asyncio.create_task(bot.delete_after([notice, sent_msg]))
         else:
             items = await dB.get_store_items(msg_id)
             if items:
-                notice = await event.reply(
-                    "⚠️ **Important Notice:**\n\n```These files will be automatically deleted after 10 minutes.```\n__Please save or forward them immediately.__"
-                )
-                sent_messages = [notice]
                 for id in items:
                     msg = await bot.get_messages(Var.CLOUD_CHANNEL, ids=id)
                     if msg:
-                        sent = await event.reply(msg.text, file=msg.media)
-                        sent_messages.append(sent)
-                asyncio.create_task(bot.delete_after(sent_messages))
+                       await event.reply(file=[i for i in msg])
     else:
         if event.sender_id == Var.OWNER:
             return await xnx.edit(
-                "** <                ADMIN PANEL                 > **",
+                "__Browse Admin Options:__",
                 buttons=admin.admin_panel(),
             )
         await event.reply(
